search for users and channels working
